requirements.txt
edgartools
pandas
pyyaml
requests
openai
tqdm
